groupai/requirements.txt
python-telegram-bot[rate-limiter]==21.1.1
PyYAML==6.0
python-dotenv==0.21.0
charade==1.0.3
openai==1.51.1
chromadb==0.5.11
pydub==0.25.1
python-docx==1.1.2
pdfplumber==0.11.4
PyMuPDF==1.24.11
pandas==1.5.3